requirements.txt
streamlit
pandas
yfinance
plotly
python-dateutil
lxml
